# make runtime consolidation about 20x faster