--- a/metric_tank/dataprocessor_test.go
+++ b/metric_tank/dataprocessor_test.go
@@ -3,6 +3,7 @@ package main
 import (
 	"github.com/raintank/raintank-metric/metric_tank/consolidation"
 	"math"
+	"math/rand"
 	"testing"
 )
 
@@ -801,6 +802,140 @@ func TestAlignRequests(t *testing.T) {
 			}
 		}
 	}
+}
+
+func randFloats(num int) []Point {
+	// let's just do the "odd" case, since the non-odd will be sufficiently close
+	num += 1
+	ret := make([]Point, num)
+	for i := 0; i < num; i++ {
+		ret[i] = Point{rand.Float64(), uint32(i)}
+	}
+	return ret
+}
+
+func randFloatsWithNulls(num int) []Point {
+	// let's just do the "odd" case, since the non-odd will be sufficiently close
+	num += 1
+	ret := make([]Point, num)
+	for i := 0; i < num; i++ {
+		if i%2 == 0 {
+			ret[i] = Point{math.NaN(), uint32(i)}
+		} else {
+			ret[i] = Point{rand.Float64(), uint32(i)}
+		}
+	}
+	return ret
+}
+
+// each "operation" (b.N) is an extra input point
+func BenchmarkConsolidateAvgRand1(b *testing.B) {
+	benchmarkConsolidate(randFloats(b.N), 1, consolidation.Avg, b)
+}
+func BenchmarkConsolidateAvgRandWithNulls1(b *testing.B) {
+	benchmarkConsolidate(randFloatsWithNulls(b.N), 1, consolidation.Avg, b)
+}
+func BenchmarkConsolidateAvgRand2(b *testing.B) {
+	benchmarkConsolidate(randFloats(b.N), 2, consolidation.Avg, b)
+}
+func BenchmarkConsolidateAvgRandWithNulls2(b *testing.B) {
+	benchmarkConsolidate(randFloatsWithNulls(b.N), 2, consolidation.Avg, b)
+}
+func BenchmarkConsolidateAvgRand25(b *testing.B) {
+	benchmarkConsolidate(randFloats(b.N), 25, consolidation.Avg, b)
+}
+func BenchmarkConsolidateAvgRandWithNulls25(b *testing.B) {
+	benchmarkConsolidate(randFloatsWithNulls(b.N), 2, consolidation.Avg, b)
+}
+func BenchmarkConsolidateAvgRand1000(b *testing.B) {
+	benchmarkConsolidate(randFloats(b.N), 1000, consolidation.Avg, b)
+}
+func BenchmarkConsolidateAvgRandWithNulls1000(b *testing.B) {
+	benchmarkConsolidate(randFloatsWithNulls(b.N), 1000, consolidation.Avg, b)
+}
+
+func BenchmarkConsolidateMinRand1(b *testing.B) {
+	benchmarkConsolidate(randFloats(b.N), 1, consolidation.Min, b)
+}
+func BenchmarkConsolidateMinRandWithNulls1(b *testing.B) {
+	benchmarkConsolidate(randFloatsWithNulls(b.N), 1, consolidation.Min, b)
+}
+func BenchmarkConsolidateMinRand2(b *testing.B) {
+	benchmarkConsolidate(randFloats(b.N), 2, consolidation.Min, b)
+}
+func BenchmarkConsolidateMinRandWithNulls2(b *testing.B) {
+	benchmarkConsolidate(randFloatsWithNulls(b.N), 2, consolidation.Min, b)
+}
+func BenchmarkConsolidateMinRand25(b *testing.B) {
+	benchmarkConsolidate(randFloats(b.N), 25, consolidation.Min, b)
+}
+func BenchmarkConsolidateMinRandWithNulls25(b *testing.B) {
+	benchmarkConsolidate(randFloatsWithNulls(b.N), 25, consolidation.Min, b)
+}
+func BenchmarkConsolidateMinRand1000(b *testing.B) {
+	benchmarkConsolidate(randFloats(b.N), 1000, consolidation.Min, b)
+}
+func BenchmarkConsolidateMinRandWithNulls1000(b *testing.B) {
+	benchmarkConsolidate(randFloatsWithNulls(b.N), 1000, consolidation.Min, b)
+}
+
+func BenchmarkConsolidateMaxRand1(b *testing.B) {
+	benchmarkConsolidate(randFloats(b.N), 1, consolidation.Max, b)
+}
+func BenchmarkConsolidateMaxRandWithNulls1(b *testing.B) {
+	benchmarkConsolidate(randFloatsWithNulls(b.N), 1, consolidation.Max, b)
+}
+func BenchmarkConsolidateMaxRand2(b *testing.B) {
+	benchmarkConsolidate(randFloats(b.N), 2, consolidation.Max, b)
+}
+func BenchmarkConsolidateMaxRandWithNulls2(b *testing.B) {
+	benchmarkConsolidate(randFloatsWithNulls(b.N), 2, consolidation.Max, b)
+}
+func BenchmarkConsolidateMaxRand25(b *testing.B) {
+	benchmarkConsolidate(randFloats(b.N), 25, consolidation.Max, b)
+}
+func BenchmarkConsolidateMaxRandWithNulls25(b *testing.B) {
+	benchmarkConsolidate(randFloatsWithNulls(b.N), 25, consolidation.Max, b)
+}
+func BenchmarkConsolidateMaxRand1000(b *testing.B) {
+	benchmarkConsolidate(randFloats(b.N), 1000, consolidation.Max, b)
+}
+func BenchmarkConsolidateMaxRandWithNulls1000(b *testing.B) {
+	benchmarkConsolidate(randFloatsWithNulls(b.N), 1000, consolidation.Max, b)
+}
+
+func BenchmarkConsolidateSumRand1(b *testing.B) {
+	benchmarkConsolidate(randFloats(b.N), 1, consolidation.Sum, b)
+}
+func BenchmarkConsolidateSumRandWithNulls1(b *testing.B) {
+	benchmarkConsolidate(randFloatsWithNulls(b.N), 1, consolidation.Sum, b)
+}
+func BenchmarkConsolidateSumRand2(b *testing.B) {
+	benchmarkConsolidate(randFloats(b.N), 2, consolidation.Sum, b)
+}
+func BenchmarkConsolidateSumRandWithNulls2(b *testing.B) {
+	benchmarkConsolidate(randFloatsWithNulls(b.N), 2, consolidation.Sum, b)
+}
+func BenchmarkConsolidateSumRand25(b *testing.B) {
+	benchmarkConsolidate(randFloats(b.N), 25, consolidation.Sum, b)
+}
+func BenchmarkConsolidateSumRandWithNulls25(b *testing.B) {
+	benchmarkConsolidate(randFloatsWithNulls(b.N), 25, consolidation.Sum, b)
+}
+func BenchmarkConsolidateSumRand1000(b *testing.B) {
+	benchmarkConsolidate(randFloats(b.N), 1000, consolidation.Sum, b)
+}
+func BenchmarkConsolidateSumRandWithNulls1000(b *testing.B) {
+	benchmarkConsolidate(randFloatsWithNulls(b.N), 1000, consolidation.Sum, b)
+}
+
+var dummy []Point
+
+func benchmarkConsolidate(in []Point, aggNum uint32, consolidator consolidation.Consolidator, b *testing.B) {
+	b.ResetTimer()
+
+	ret := consolidate(in, aggNum, consolidator)
+	dummy = ret
 }
 
 var result []Req