--- a/schema/point.go
+++ b/schema/point.go
@@ -1,0 +1,6 @@
+package schema
+
+type Point struct {
+	Val float64
+	Ts  uint32
+}