```
benchmark                                               old ns/op       new ns/op       delta
BenchmarkConsolidateAvgRand1M_1-8                       182598742       8204137         -95.51%
BenchmarkConsolidateAvgRandWithNulls1M_1-8              139330755       8003439         -94.26%
BenchmarkConsolidateAvgRand1M_2-8                       143220273       4650865         -96.75%
BenchmarkConsolidateAvgRandWithNulls1M_2-8              98254475        4295599         -95.63%
BenchmarkConsolidateAvgRand1M_25-8                      61812960        1991016         -96.78%
BenchmarkConsolidateAvgRandWithNulls1M_25-8             49045875        2120413         -95.68%
BenchmarkConsolidateAvgRand1M_100-8                     46049393        2048007         -95.55%
BenchmarkConsolidateAvgRandWithNulls1M_100-8            35081646        1930699         -94.50%
BenchmarkConsolidateMinRand1M_1-8                       100985317       8706482         -91.38%
BenchmarkConsolidateMinRandWithNulls1M_1-8              71410033        8006405         -88.79%
BenchmarkConsolidateMinRand1M_2-8                       82802109        6831486         -91.75%
BenchmarkConsolidateMinRandWithNulls1M_2-8              58118902        4607662         -92.07%
BenchmarkConsolidateMinRand1M_25-8                      39843635        2820777         -92.92%
BenchmarkConsolidateMinRandWithNulls1M_25-8             26662616        2699100         -89.88%
BenchmarkConsolidateMinRand1M_100-8                     29310478        2308417         -92.12%
BenchmarkConsolidateMinRandWithNulls1M_100-8            18692594        2177030         -88.35%
BenchmarkConsolidateMaxRand1M_1-8                       98419876        8613543         -91.25%
BenchmarkConsolidateMaxRandWithNulls1M_1-8              71995637        8668712         -87.96%
BenchmarkConsolidateMaxRand1M_2-8                       83930212        7321626         -91.28%
BenchmarkConsolidateMaxRandWithNulls1M_2-8              52415588        5015605         -90.43%
BenchmarkConsolidateMaxRand1M_25-8                      39580950        2945151         -92.56%
BenchmarkConsolidateMaxRandWithNulls1M_25-8             26219169        2546342         -90.29%
BenchmarkConsolidateMaxRand1M_100-8                     28796999        2271480         -92.11%
BenchmarkConsolidateMaxRandWithNulls1M_100-8            18604390        2106965         -88.67%
BenchmarkConsolidateSumRand1M_1-8                       99188202        6995902         -92.95%
BenchmarkConsolidateSumRandWithNulls1M_1-8              71143768        7460804         -89.51%
BenchmarkConsolidateSumRand1M_2-8                       80475413        4136470         -94.86%
BenchmarkConsolidateSumRandWithNulls1M_2-8              51512634        4382817         -91.49%
BenchmarkConsolidateSumRand1M_25-8                      37280725        1634229         -95.62%
BenchmarkConsolidateSumRandWithNulls1M_25-8             25140374        1774661         -92.94%
BenchmarkConsolidateSumRand1M_100-8                     28309718        1849072         -93.47%
BenchmarkConsolidateSumRandWithNulls1M_100-8            18261723        1968354         -89.22%

benchmark                                               old allocs     new allocs     delta
BenchmarkConsolidateAvgRand1M_1-8                       2000033        0              -100.00%
BenchmarkConsolidateAvgRandWithNulls1M_1-8              1000027        1              -100.00%
BenchmarkConsolidateAvgRand1M_2-8                       2000025        0              -100.00%
BenchmarkConsolidateAvgRandWithNulls1M_2-8              1000020        0              -100.00%
BenchmarkConsolidateAvgRand1M_25-8                      480021         0              -100.00%
BenchmarkConsolidateAvgRandWithNulls1M_25-8             400013         0              -100.00%
BenchmarkConsolidateAvgRand1M_100-8                     160017         0              -100.00%
BenchmarkConsolidateAvgRandWithNulls1M_100-8            140012         0              -100.00%
BenchmarkConsolidateMinRand1M_1-8                       1000021        2              -100.00%
BenchmarkConsolidateMinRandWithNulls1M_1-8              500013         0              -100.00%
BenchmarkConsolidateMinRand1M_2-8                       1000011        1              -100.00%
BenchmarkConsolidateMinRandWithNulls1M_2-8              500015         1              -100.00%
BenchmarkConsolidateMinRand1M_25-8                      240013         0              -100.00%
BenchmarkConsolidateMinRandWithNulls1M_25-8             200005         0              -100.00%
BenchmarkConsolidateMinRand1M_100-8                     80010          0              -100.00%
BenchmarkConsolidateMinRandWithNulls1M_100-8            70004          0              -100.00%
BenchmarkConsolidateMaxRand1M_1-8                       1000015        0              -100.00%
BenchmarkConsolidateMaxRandWithNulls1M_1-8              500013         0              -100.00%
BenchmarkConsolidateMaxRand1M_2-8                       1000015        1              -100.00%
BenchmarkConsolidateMaxRandWithNulls1M_2-8              500008         1              -100.00%
BenchmarkConsolidateMaxRand1M_25-8                      240013         0              -100.00%
BenchmarkConsolidateMaxRandWithNulls1M_25-8             200004         0              -100.00%
BenchmarkConsolidateMaxRand1M_100-8                     80010          0              -100.00%
BenchmarkConsolidateMaxRandWithNulls1M_100-8            70004          0              -100.00%
BenchmarkConsolidateSumRand1M_1-8                       1000016        1              -100.00%
BenchmarkConsolidateSumRandWithNulls1M_1-8              500006         0              -100.00%
BenchmarkConsolidateSumRand1M_2-8                       1000016        0              -100.00%
BenchmarkConsolidateSumRandWithNulls1M_2-8              500007         0              -100.00%
BenchmarkConsolidateSumRand1M_25-8                      240013         0              -100.00%
BenchmarkConsolidateSumRandWithNulls1M_25-8             200006         0              -100.00%
BenchmarkConsolidateSumRand1M_100-8                     80010          0              -100.00%
BenchmarkConsolidateSumRandWithNulls1M_100-8            70004          0              -100.00%

benchmark                                               old bytes     new bytes     delta
BenchmarkConsolidateAvgRand1M_1-8                       32008856      63            -100.00%
BenchmarkConsolidateAvgRandWithNulls1M_1-8              24008524      65            -100.00%
BenchmarkConsolidateAvgRand1M_2-8                       32004969      34            -100.00%
BenchmarkConsolidateAvgRandWithNulls1M_2-8              16004515      41            -100.00%
BenchmarkConsolidateAvgRand1M_25-8                      40968056      28            -100.00%
BenchmarkConsolidateAvgRandWithNulls1M_25-8             20487804      30            -100.00%
BenchmarkConsolidateAvgRand1M_100-8                     40965183      24            -100.00%
BenchmarkConsolidateAvgRandWithNulls1M_100-8            20485134      18            -100.00%
BenchmarkConsolidateMinRand1M_1-8                       24008111      131           -100.00%
BenchmarkConsolidateMinRandWithNulls1M_1-8              20007932      58            -100.00%
BenchmarkConsolidateMinRand1M_2-8                       20004096      85            -100.00%
BenchmarkConsolidateMinRandWithNulls1M_2-8              12004292      79            -100.00%
BenchmarkConsolidateMinRand1M_25-8                      20807815      20            -100.00%
BenchmarkConsolidateMinRandWithNulls1M_25-8             10567571      16            -100.00%
BenchmarkConsolidateMinRand1M_100-8                     20565032      21            -100.00%
BenchmarkConsolidateMinRandWithNulls1M_100-8            10324892      26            -100.00%
BenchmarkConsolidateMaxRand1M_1-8                       24007893      51            -100.00%
BenchmarkConsolidateMaxRandWithNulls1M_1-8              20007945      39            -100.00%
BenchmarkConsolidateMaxRand1M_2-8                       20004347      116           -100.00%
BenchmarkConsolidateMaxRandWithNulls1M_2-8              12004046      96            -100.00%
BenchmarkConsolidateMaxRand1M_25-8                      20807777      30            -100.00%
BenchmarkConsolidateMaxRandWithNulls1M_25-8             10567560      16            -100.00%
BenchmarkConsolidateMaxRand1M_100-8                     20565026      32            -100.00%
BenchmarkConsolidateMaxRandWithNulls1M_100-8            10324899      30            -100.00%
BenchmarkConsolidateSumRand1M_1-8                       24007824      65            -100.00%
BenchmarkConsolidateSumRandWithNulls1M_1-8              20007470      60            -100.00%
BenchmarkConsolidateSumRand1M_2-8                       20004342      17            -100.00%
BenchmarkConsolidateSumRandWithNulls1M_2-8              12003940      31            -100.00%
BenchmarkConsolidateSumRand1M_25-8                      20807852      22            -100.00%
BenchmarkConsolidateSumRandWithNulls1M_25-8             10567662      17            -100.00%
BenchmarkConsolidateSumRand1M_100-8                     20565033      25            -100.00%
BenchmarkConsolidateSumRandWithNulls1M_100-8            10324926      14            -100.00%
```