@Dieterbe 2016-03-08
now also with a faster `fix()`. could only shave off 18% though.