@Dieterbe 2016-03-08
besides consolidate and fix having a good set of unit tests, i also loaded 24h of data using fake_data_to_nsq, played with different timescales, and confirmed that the consolidations looks good across a range of rollup ratios (using a chart for each of avg/sum/min/max, it's easy to tell it picks the right values)
since consolidate now modifies slices in place and reuses the input for output, i also did the same with a `-race` enabled build , and ran a vegeta http load of 24h of data at 10/s in the background

```
cat data-test| vegeta report | grep -v 'connection reset'
Requests      [total, rate]            2816, 10.00
Duration      [total, attack, wait]    4m41.500548601s, 4m41.49999993s, 548.671µs
Latencies     [mean, 50, 95, 99, max]  12.309940951s, 3.924787396s, 30.001371159s, 30.003535594s, 30.210600178s
Bytes In      [total, mean]            19478534, 6917.09
Bytes Out     [total, mean]            0, 0.00
Success       [ratio]                  9.73%
Status Codes  [code:count]             200:274  0:2542  
Error Set:
Get http://localhost:6063/get?to=1457408230&from=1456803430&target=1.e41d5e3a6b4e54154684c60177a65641&target=1.f27d1c8a4c4616906bae83ce75f125f0&maxDataPoints=1920: EOF
Get http://localhost:6063/get?to=1457408230&from=1456803430&target=consolidateBy%281.e41d5e3a6b4e54154684c60177a65641%2C%27sum%27%29&target=consolidateBy%281.f27d1c8a4c4616906bae83ce75f125f0%2C%27sum%27%29&maxDataPoints=1920: EOF
```

the problem is cassandra on my laptop can't do all that load so it started getting slow and timing out toward the end, but anyway the main goal of this test was asserting functionality under concurrency, there was no mention of a data race in the log.